requirements.txt
line-bot-sdk==3.14.1
openai==1.77.0
flask==3.1.0
python-dotenv==1.1.0
gunicorn==25.3.0
